#44 The Collaborations map and Individual search doesn't work as expected
The error indicated that the code was trying to access a column named title, which does not exist in the uploaded CSV file. Because of this, the app failed to render outputs in sections where the "title" column was used. So once the "title" column was commented out the map and other aspects of the app started to work.